api/requirements.txt
amqp==5.0.6
bidict==0.21.2
billiard==3.6.4.0
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Flask==1.1.2
Flask-SocketIO==5.0.1
itsdangerous==1.1.0
Jinja2==2.11.3
kombu==5.0.2
librabbitmq==2.0.0
MarkupSafe==1.1.1
mongoengine==0.23.0
prompt-toolkit==3.0.18
pymongo==3.11.3
python-dotenv==0.17.1
python-engineio==4.1.0
python-socketio==5.2.1
pytz==2021.1
six==1.15.0
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
